Get_pom_client Consolidated apis
