+++ server/data/curious/curiousApi.test.ts
