+++ server/data/curious/curiousApi.ts
