+++ server/data/curious/types/Enums.ts
