+++ server/data/curious/types/Types.ts
