+++ server/data/curious/types/index.d.ts
